backend/requirements.txt
fastapi>=0.110.0
uvicorn>=0.28.0
pydantic>=2.7.4,<3.0.0
pydantic-core>=2.33.2
python-dotenv>=1.0.1
boto3>=1.34.0
openai>=1.24.0
pinecone-client>=6.0.0
supabase>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.0.0
python-multipart>=0.0.9
httpx>=0.24.0
stripe>=8.0.0
pytest>=8.0.0
PyJWT>=2.8.0
pytz>=2024.1
pypdf>=4.0.0
PyPDF2>=3.0.0
python-docx>=1.0.0
email-validator>=2.0.0
requests>=2.25.0
click>=8.1.0
langchain>=0.3.24,<0.4.0
langchain-core>=0.3.56,<0.4.0
langchain-openai>=0.3.0
langchain-pinecone>=0.2.5
langchain-community>=0.3.23
langgraph>=0.3.0
langgraph-prebuilt>=0.1.8 